--- src/app/about/about.component.html
@@ -1,6 +1,0 @@
-<div class="container">
-  Docks-UI is the frontend for Docks; A docker swarm manager-er
-</div>
-<div class="container">
-  <button type="button" class="btn btn-primary" (click)="reveal()">Show Containers</button>
-</div>
