+++ src/app/about/about.component.ts
@@ -12,12 +12,4 @@ export class AboutComponent implements OnInit {
 
   ngOnInit() {
   }
-
-  public reveal() {
-    this.http.get('http://localhost:8080/docker/containers/json', { responseType: 'json' }).subscribe( json => {
-        for ( let i = 0; i < Object.keys(json).length; i++) {
-          alert('Container ' + <number>(i + 1) + ': ' + json[i]['Id'].slice(0, 10));
-        }
-    });
-  }
 }
