+++ src/app/app.module.ts
@@ -3,10 +3,10 @@ import { NgModule } from '@angular/core';
 import { AppRoutingModule } from './app-routing.module';
 import { HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { MaterialModule } from './material.module';
 import { AppComponent } from './app.component';
 import { AboutComponent } from './about/about.component';
-import {MatButtonModule} from '@angular/material';
-
+import { FormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [
@@ -18,7 +18,8 @@ import {MatButtonModule} from '@angular/material';
       AppRoutingModule,
       HttpClientModule,
       BrowserAnimationsModule,
-      MatButtonModule,
+      MaterialModule,
+      FormsModule,
   ],
   providers: [],
   bootstrap: [AppComponent]
